fix: Renamed Tracking.js to MetricsHelper.js
